Abstract instruction dispatch into a trait.
The nfa and backtracking implementations were reimplementing logic for
dispatching an instruction against a position in the input stream. I
factored this out into a `Dispatch` trait which abstracts the common
logic. This actually makes the code a little bigger, but it improves
orthogonality and will make it easier to switch to a bytecode (or
whatever), if desired.